pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "MEDS_tabularization"
version = "0.0.1"
authors = [
  { name="Matthew McDermott", email="mattmcdermott8@gmail.com" },
  { name="Nassim Oufattole", email="noufattole@gmail.com" },
  { name="Teya Bergamaschi", email="teyabergamaschi@gmail.com" },
]
description = "Scalable Tabularization of MEDS format Time-Series data"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = ["polars", "pyarrow", "loguru", "hydra-core", "numpy", "scipy", "pandas", "tqdm", "xgboost", "scikit-learn", "hydra-optuna-sweeper", "hydra-joblib-launcher"]

[project.scripts]
meds-tab-describe = "MEDS_tabular_automl.scripts.describe_codes:main"
meds_tab-tabularize-static = "MEDS_tabular_automl.scripts.tabularize_static:main"
meds_tab-tabularize-time-series = "MEDS_tabular_automl.scripts.tabularize_time_series:main"
meds_tab-cache-task = "MEDS_tabular_automl.scripts.cache_task:main"
meds_tab-xgboost = "MEDS_tabular_automl.scripts.launch_xgboost:main"

[project.optional-dependencies]
dev = ["pre-commit"]
tests = ["pytest", "pytest-cov", "rootutils"]
profiling = ["mprofile", "matplotlib"]

[project.urls]
Homepage = "https://github.com/mmcdermott/MEDS_Tabular_AutoML"
Issues = "https://github.com/mmcdermott/MEDS_Tabular_AutoML/issues"
